pyproject.toml
[project]
name = "hexlet-code"
version = "0.1.0"
description = "Page Analyzer"
requires-python = ">=3.10"

dependencies = [
    "flask",
    "python-dotenv",
    "gunicorn",
    "psycopg[binary]",
    "validators"
]

[tool.uv]
dev-dependencies = [
    "flake8"
]

